Renamed quickstart.py and the associated import in views.py, added basic readme
